Won another 0.1 of second on double -> float

+++ src/cluster/AbstractMetric.h
@@ -16,7 +16,7 @@
 class AbstractMetric {
 public:
 
-    virtual double distance(Object &o1, Object &o2, bool bUseIntegratedPrediction = false) = 0;
+    virtual float distance(Object &o1, Object &o2, bool bUseIntegratedPrediction = false) = 0;
     virtual Object* center(DataContainer *pContainer, \
         list<int> &indices) = 0;
     virtual void predictMissingData(DataContainer *pContainer) = 0;
