+++ src/main/Regression.cpp
@@ -1,7 +1,7 @@
 #include "Regression.h"
 #include <cmath>
 
-double Regression::predict(double **matrix, double *known, double *pred, int nRows, int nCols) {
+float Regression::predict(float **matrix, float *known, float *pred, int nRows, int nCols) {
 
 	return 0;
 }
